pyproject.toml
[project]
name = "fastapi-hexagonal-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.126.0",
    "passlib>=1.7.4",
    "pre-commit>=4.5.1",
    "psycopg>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "ruff>=0.14.10",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
